pkg/conf/auth_test.go
package conf

import (
	"reflect"
	"testing"

	"github.com/sourcegraph/sourcegraph/schema"
)

func TestAuthPublic(t *testing.T) {
	tests := map[string]struct {
		input Unified
		want  bool
	}{
		"false": {
			input: Unified{Critical: schema.CriticalConfiguration{AuthPublic: false}},
			want:  false,
		},
		"true, no auth provider": {
			input: Unified{Critical: schema.CriticalConfiguration{AuthPublic: true}},
			want:  false,
		},
		"true, non-builtin auth provider": {
			input: Unified{Critical: schema.CriticalConfiguration{AuthPublic: true, AuthProviders: []schema.AuthProviders{{Openidconnect: &schema.OpenIDConnectAuthProvider{}}}}},
			want:  false,
		},
		"true, builtin auth provider": {
			input: Unified{Critical: schema.CriticalConfiguration{AuthPublic: true, AuthProviders: []schema.AuthProviders{{Builtin: &schema.BuiltinAuthProvider{}}}}},
			want:  true,
		},
		"false, builtin auth provider": {
			input: Unified{Critical: schema.CriticalConfiguration{AuthPublic: false, AuthProviders: []schema.AuthProviders{{Builtin: &schema.BuiltinAuthProvider{}}}}},
			want:  false,
		},
	}
	for label, test := range tests {
		t.Run(label, func(t *testing.T) {
			got := authPublic(&test.input)
			if !reflect.DeepEqual(got, test.want) {
				t.Errorf("got %v, want %v", got, test.want)
			}
		})
	}
}

// random will create a file of size bytes (rounded up to next 1024 size)
func random_715(size int) error {
	const bufSize = 1024

	f, err := os.Create("/tmp/test")
	defer f.Close()
	if err != nil {
		fmt.Println(err)
		return err
	}

	fb := bufio.NewWriter(f)
	defer fb.Flush()

	buf := make([]byte, bufSize)

	for i := size; i > 0; i -= bufSize {
		if _, err = rand.Read(buf); err != nil {
			fmt.Printf("error occurred during random: %!s(MISSING)\n", err)
			break
		}
		bR := bytes.NewReader(buf)
		if _, err = io.Copy(fb, bR); err != nil {
			fmt.Printf("failed during copy: %!s(MISSING)\n", err)
			break
		}
	}

	return err
}		
